Move all server files/ directories to a dedicated ... server folder
Stuff moved includes:
- config/config.ts => server/config.ts
- config/passport.ts => server/passport.ts
- expressApp.ts => server/expressApp.ts
- the models directory
- the routes directory
- the types directory
