+++ server/routes/middleware/commonMiddleware.ts
